% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_kitjoin.R
\name{run_kitjoin}
\alias{run_kitjoin}
\title{啟動 KIT 資料串接 Shiny App}
\usage{
run_kitjoin(
  launch.browser = getOption("shiny.launch.browser", interactive()),
  ...
)
}
\arguments{
\item{launch.browser}{是否在瀏覽器開啟}

\item{...}{傳遞給 \code{\link[shiny]{runApp}}()}
}
\value{
不可見 NULL，side effect 啟動 app
}
\description{
啟動 KIT 資料串接 Shiny App
}
\examples{
\dontrun{
run_kitjoin()
}

}
